refactor: 기존 PageSize constants 디렉토리로 변경
